delete files

+++ app/Services/ImageService.php
@@ -8,11 +8,15 @@ use App\Http\Requests\Profile\SaveCongratulationRequest;
 use App\Http\Requests\SaveReviewRequest;
 use App\Models\Banner;
 use App\Models\Logo;
+use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Intervention\Image\ImageManagerStatic as Image;
 
 class ImageService {
+    const UPLOAD_PATH = 'images/upload_images/';
+    const RESIZE_PATH = 'images/resize_images/';
+
     /**
      * @param SaveCongratulationRequest|SaveReviewRequest|Request $request
      * @param string $path
@@ -24,16 +28,16 @@ class ImageService {
         $originalName = $file->getClientOriginalName();
         $originalName = str_replace(' ', '', $originalName);
         $fileName = time().'_'.$originalName;
-        Storage::disk('public')->putFileAs('images/upload_images/' . $path . DIRECTORY_SEPARATOR, $file, $fileName);
-        Storage::disk('public')->makeDirectory('images/resize_images/' . $path . DIRECTORY_SEPARATOR);
+        Storage::disk('public')->putFileAs(self::UPLOAD_PATH . $path . DIRECTORY_SEPARATOR, $file, $fileName);
+        Storage::disk('public')->makeDirectory(self::RESIZE_PATH . $path . DIRECTORY_SEPARATOR);
         Image::make($file->getRealPath())
             ->resize(150, 150, function($img){
                 $img->aspectRatio();
             })->resizeCanvas(150, 150, 'center', false, '#f2f2f2')
-            ->save(storage_path('app/public/images/resize_images/' . $path . DIRECTORY_SEPARATOR . $fileName));
+            ->save(storage_path('app/public/' . self::RESIZE_PATH . $path . DIRECTORY_SEPARATOR . $fileName));
 
         return [
-            'src' => 'images/upload_images/' . $path . DIRECTORY_SEPARATOR . $fileName,
+            'src' => self::UPLOAD_PATH . $path . DIRECTORY_SEPARATOR . $fileName,
             'original_name' => $originalName,
             'name' => $fileName
         ];
@@ -42,25 +46,35 @@ class ImageService {
     /**
      * @param SaveCongratulationRequest|SaveReviewRequest|Request $request
      * @param $item
-     * @param string $path
      *
      * @return array
      */
-    public static function updateImage($request, $item, $path = 'reviews'){
+    public static function updateImage($request, $item){
         $imageInfo = self::uploadImage($request);
         if($item->image){
-            $filePath = 'images/upload_images/' . $path . DIRECTORY_SEPARATOR . $item->image->name;
-            $is_exist = Storage::disk('public')->exists($filePath);
-            if($is_exist){
-                Storage::disk('public')->delete($filePath);
-                $resizeFilePath = 'images/resize_images/' . $path . DIRECTORY_SEPARATOR . $item->image->name;
-                if(Storage::disk('public')->delete($resizeFilePath)){
-                    Storage::disk('public')->delete($resizeFilePath);
-                }
-            }
+            self::deleteImage($item);
         }
 
         return $imageInfo;
+    }
+
+    /**
+     * @param string $path
+     * @param Model $item
+     */
+    public static function deleteImage(Model $item, string $path = 'reviews')
+    {
+
+        $filePath = self::UPLOAD_PATH . $path . DIRECTORY_SEPARATOR . optional($item->image)->name;
+        $is_exist = Storage::disk('public')->exists($filePath);
+        if($is_exist && $item->image){
+            Storage::disk('public')->delete($filePath);
+            $resizeFilePath = self::RESIZE_PATH . $path . DIRECTORY_SEPARATOR . optional($item->image)->name;
+            if(Storage::disk('public')->delete($resizeFilePath)){
+                Storage::disk('public')->delete($resizeFilePath);
+            }
+            $item->image->delete();
+        }
     }
 
     public static function uploadBanner($request){
